chore(vrt): Refactor visual regression tests for stability and code reduction (#9657)
