+++ tests/playwright/vrt-hr-components.spec.ts
@@ -2,7 +2,6 @@ import { type BrowserContext, type Page, expect } from '@playwright/test'
 import { test } from './fixtures'
 import {
   getDynamicContentMasks,
-  getHrMasks,
   setupVisualRegressionTest,
   mockMultipleHrDevices,
   resetServerState,
@@ -89,10 +88,7 @@ test.describe('Visual Regression Tests', () => {
 
       await takeScreenshot(dashboard, 'dashboard-with-hr-data.png', {
         maxDiffPixelRatio: 0.15,
-        mask: [
-          ...getDynamicContentMasks(dashboardPage),
-          ...getHrMasks(dashboardPage),
-        ],
+        mask: [...getDynamicContentMasks(dashboardPage)],
       })
     })
 
@@ -134,10 +130,7 @@ test.describe('Visual Regression Tests', () => {
       const dashboard = dashboardPage.getByTestId('dashboard')
       await takeScreenshot(dashboard, 'dashboard-with-2-hr-devices.png', {
         maxDiffPixelRatio: 0.15,
-        mask: [
-          ...getDynamicContentMasks(dashboardPage),
-          ...getHrMasks(dashboardPage),
-        ],
+        mask: [...getDynamicContentMasks(dashboardPage)],
       })
     })
 
@@ -159,10 +152,7 @@ test.describe('Visual Regression Tests', () => {
         const dashboard = dashboardPage.getByTestId('dashboard')
         await takeScreenshot(dashboard, `dashboard-hr-zone-${zone}.png`, {
           maxDiffPixelRatio: 0.1,
-          mask: [
-            ...getDynamicContentMasks(dashboardPage),
-            ...getHrMasks(dashboardPage),
-          ],
+          mask: [...getDynamicContentMasks(dashboardPage)],
         })
       })
     }
@@ -173,10 +163,7 @@ test.describe('Visual Regression Tests', () => {
       const dashboard = dashboardPage.getByTestId('dashboard')
       await takeScreenshot(dashboard, 'dashboard-hr-disconnected.png', {
         maxDiffPixelRatio: 0.1,
-        mask: [
-          ...getDynamicContentMasks(dashboardPage),
-          ...getHrMasks(dashboardPage),
-        ],
+        mask: [...getDynamicContentMasks(dashboardPage)],
       })
     })
   })
